bleh
git-svn-id: svn+ssh://svn.code.sf.net/p/snowstorm/code/trunk/openglad@284 636a839e-d62e-0410-834d-a1d0aacdb804

+++ video.h
@@ -68,7 +68,7 @@ class video
          void clear_ints();
          void restore_ints();
 
-	void get_pixel(int x, int y, int *r, int *g, int *b);
+	void get_pixel(int x, int y, Uint8 *r, Uint8 *g, Uint8 *b);
 	int get_pixel(int x, int y, int *index);
 	int get_pixel(int offset);
